Update clients controller to use cache
